Checkpoint #8.5: Updates and code-maintenance
  Vim/Neovim:
  c Finish migration: null-ls -> none-ls -> nvim-lint + conform.nvim
  + Add nvim-lint and conform.nvim plugins and their configurations.
  - Remove 'null-ls' and 'none-ls' custom settings.
  +/c Adapt to the changes above and rewrite a bit 'lspconfig.lua'.
  c Adapt 'lualine.nvim' to the changes above.
  c Run 'telescope' temporal layout options.
    This helps to avoid conflicts when running plugins w/ their own layout
    settings for pickers/previewers (like 'actions-preview.nvim').
  + Add "resume" mapping to continue browsing telescope search results.

  + `<Leader>nu` now toggles also indentation lines of 'indent-blankline.nvim'.
  + Return toggle to compare two buffers with `:diffthis`.
  c Format files that were changed.

  Shell:
  c Change `v` and `vv` aliases functionality a bit.
  + `tarz` now saves to the current folder (not to the target's parent).
  * `bak` now doesn't hold on removing intermediate file/directory.

This is a commit fraction #5.
It intends to finish this branch of development.
There might be another commit #6 that fixes some minor bugs.

NOTE: Checkpoint #9 is not finished.
NOTE: Checkpoint #10 is not finished.

We got back to 8-th checkpoint :)
But we will get back to the rest soon.

Next commits (or the last fraction) may be about:
- backup and laptop migration pipelines.
- optimizing loading of plugins with `lazy = true` option.
- updating README

+++ _impl/control.sh
@@ -336,8 +336,11 @@ control::uninstall () {
 
   rm -f ~/.condarc
   rm -f ~/.tmux.conf
-  rm -rf "$XDG_CONFIG_HOME/nvim"
-  rm -f "$XDG_CONFIG_HOME/tmux/.tmux.conf"
+  if [[ -n "$XDG_CONFIG_HOME" ]]; then
+    ## Unlikely there will be /nvim and /tmux dirs, nevertheless..
+    rm -rf "$XDG_CONFIG_HOME/nvim"
+    rm -f "$XDG_CONFIG_HOME/tmux/.tmux.conf"
+  fi
 
   if grep -q '^jupyter' .update-list
   then
